2010-01-20 Roman - SYS: RPM-spec files is updated for openscada_demo call and user's passwords fixing. - !!!!:SYS: The files debian/* is added for build Debian packages support. - !!:DemoDB: User call for Demo DB support is added for write to DB support. - FIX:Archive.Value: Last value archivator call is added on archive task stop for present data save. - FIX:SYS: Text encoding crash on error symbols into begin encoded text's is fixed.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1053 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tmess.cpp
@@ -170,7 +170,7 @@ string TMess::codeConv( const string &fromCH, const string &toCH, const string &
 	{
 	    mess_crit("IConv",_("Error input sequence convert: %s"),strerror(errno));
 	    mess_debug("IConv",_("Error converting from %s to %s for message part: '%s'"),
-		fromCH.c_str(),toCH.c_str(),mess.substr(vmax(mess.size()-ilen-10,0),20).c_str());
+		fromCH.c_str(),toCH.c_str(),mess.substr(vmax((int)mess.size()-(int)ilen-10,0),20).c_str());
 	    buf = mess;
 	    break;
 	}
